clean AsyncShift
